+++ elements/share/src/components/MessagesIntlProvider.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import { IntlProvider, injectIntl, InjectedIntlProps } from 'react-intl';
 import { getMessagesForLocale, LangCode } from '../util/i18n-util';
 
@@ -6,16 +6,17 @@ export interface Props {
   children: React.ReactChild;
 }
 
-class MessagesIntlProvider extends React.Component<Props & InjectedIntlProps> {
-  render() {
-    const { intl, children } = this.props;
-
-    return (
-      <IntlProvider messages={getMessagesForLocale(intl.locale as LangCode)}>
-        {children}
-      </IntlProvider>
-    );
-  }
-}
+const MessagesIntlProvider = ({
+  intl,
+  children,
+}: Props & InjectedIntlProps) => {
+  const mergedMessages = useMemo(() => {
+    return {
+      ...intl.messages,
+      ...getMessagesForLocale(intl.locale as LangCode),
+    };
+  }, [intl.messages, intl.locale]);
+  return <IntlProvider messages={mergedMessages}>{children}</IntlProvider>;
+};
 
 export default injectIntl(MessagesIntlProvider);