+++ navigation/atlassian-switcher/src/ui/primitives/index.tsx
@@ -12,3 +12,4 @@ export { default as TryLozenge } from './try-lozenge';
 export { default as FormattedMessage } from './formatted-message';
 export { default as ItemWithAvatarGroup } from './item-with-avatar-group';
 export { default as SlackIcon } from './SlackIcon';
+export { default as NotificationDot } from './notification-dot';